% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{bonholm_cor}
\alias{bonholm_cor}
\title{Bonferonni-Holm}
\usage{
bonholm_cor(p_vals, alpha = 0.05)
}
\arguments{
\item{p_vals}{Enter a vector of p-values to compute the bonferroni-holm correction on.}

\item{alpha}{The alpha level of you significance test.}
}
\value{
\item{nSig}{The number of significant results.}
\item{p_val}{The significant p-values.}
}
\description{
This function computes the bonferroni-holm correction and determines if the p-values
entered are significant or not.
}
