% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvsplit.R
\name{cvsplit}
\alias{cvsplit}
\title{Cross-Validation Splitting}
\usage{
cvsplit(train = NULL, folds = 10)
}
\arguments{
\item{train}{The dataset you wish to split.}

\item{folds}{The amount of folds you wish to produce. Note the amount of observations divided by folds must equal a whole number.}
}
\value{
\item{trainlist}{A list of the training sets}
\item{vallist}{A list of the validation sets}
}
\description{
This function shuffles the observations of a dataset, splits it into k-folds and produces a
 list of the training sets and a list of the validation sets. Furthermore
}
\examples{
data(iris)
cv_lists <- cvsplit(iris)
}
