% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{Mode}
\alias{Mode}
\title{Mode function}
\usage{
Mode(vector)
}
\arguments{
\item{vector}{Vector of the data you want to calculate the mode of.}
}
\value{
\item{mode} The value most often found in the vector.
}
\description{
This function allows you to calculate the mode of a vector.
The mode is the most often value in a vector.
}
