% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onehot_encoder.R
\name{onehot_encode}
\alias{onehot_encode}
\title{A One-Hot Encoder}
\usage{
onehot_encode(factor_vector, name)
}
\arguments{
\item{factor_vector}{Vector containing the factors you want to one-hot encode.}

\item{name}{Name of the coloumns in the output matrix.}
}
\value{
\item{one-hot_matrix} Encoded matrix for the factors.
}
\description{
This functions allows you to transform a vector into a One-Hot Encoded matrix.
Useful for nominal input data and tensorflow neural network outputs.
}
\keyword{onehot-encoding}
