% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcafuncs.R
\name{pca_biplot}
\alias{pca_biplot}
\title{Biplot of a PCA model}
\usage{
pca_biplot(model, point_id = NULL, dim = c(1, 2), min_cor = 0.3,
  biplot = TRUE)
}
\arguments{
\item{model}{The PCA model from FactoMineR.}

\item{point_id}{A vector of labels to the individual points. Ignore this if you do not want labels.}

\item{dim}{The dimensions to be plotted. Defaults to c(1,2).}

\item{min_cor}{The minimal correlation value to plot arrow. Defaults to .3.}

\item{biplot}{A boolean for whether or not to produce a biplot and therefore add correlation arrows. Defaults to TRUE}
}
\value{
\item{ind}{Data of the individual coordinates on the dimensions.}
\item{loadings}{Data of the variable loadings on the dimensions.}
\item{plot}{The ggplot object.}
}
\description{
This function creates a biplot of the PCA model. You can adjust which dimensions to plot,
aswell as the threshold for correlations to be shown. You can also forego the correlation arrows with biplot = FALSE.
}
