% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcafuncs.R
\name{pca_loading_circle}
\alias{pca_loading_circle}
\title{Loadings Circle}
\usage{
pca_loading_circle(model, dim = c(1, 2), min_cor = 0.3)
}
\arguments{
\item{model}{The PCA model from FactoMineR.}

\item{dim}{The dimensions to be plotted. Defaults to c(1,2).}

\item{min_cor}{The minimal correlation value to plot arrow. Defaults to .3.}
}
\value{
\item{loadings}{Data of the variable loadings on the dimensions.}
\item{plot}{The ggplot object.}
}
\description{
This function creates a unity circle aswell as the correlations between the original
variables and the principal components of your PCA model. You can choose which dimensions to plot,
aswell as the threshold for correlations to be shown.
}
