% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcafuncs.R
\name{pca_scree}
\alias{pca_scree}
\title{Classic Screeplot}
\usage{
pca_scree(model, type = "cpov")
}
\arguments{
\item{model}{The PCA model from FactoMineR.}

\item{type}{The type of model. This can be one of three: "eig", "pov" or "cpov". Defaults to "cpov".}
}
\description{
This function constructs a screeplot based on either the eigenvalues, percentage of variance or the cummulative percentage of variance.
this can be done by adjusting the type variable.
}
