% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{A QQ plot}
\usage{
qqplot(Y = NULL, X = NULL, type = "normal",
  ylabel = "Observed Values", xlabel = "Theoretical Quantiles")
}
\arguments{
\item{Y}{A data vector to plot on the Y axis.}

\item{X}{Additional data vector that will overwrite the normality quantiales that usually go on the X axis.
Useful to compare distributions of different vectors.}

\item{type}{The type of quantiles. Either "normal" or "t" distributed.}

\item{ylabel}{The label for the x axis.}

\item{xlabel}{The label for the y axis.}
}
\value{
\item{data}{A dataset containing the vectors used for the plot.}
\item{plot}{A ggplot object of the plot.}
}
\description{
Produces a QQ plot of your data with either normal or t-distributed quantiles.
}
\examples{
data(mtcars)
qqplot(mtcars$mpg)

data(mtcars)
qqplot(mtcars$mpg, mtcars$qsec)
}
