% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{A Root-Mean-Square Error function}
\usage{
rmse(error)
}
\arguments{
\item{error}{Vector containing the difference between model predictions and true values.}
}
\value{
\item{rms}{Root Mean Square Error.}
}
\description{
This function allows you to quickly calculate the RMSE of a error vector.
}
\examples{
model <- lm(mpg~., data = mtcars)
predictions <- predict(model, mtcars)
error <- mtcars$mpg - predictions
rms <- rmse(error)
}
\keyword{rmse}
