% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{A QQ plot}
\usage{
qqplot(data = NULL, type = "normal", ylabel = "Observed Values",
  xlabel = "Theoretical Quantiles")
}
\arguments{
\item{data}{The dataset only containing the data coloumns correlations between.}

\item{type}{The type of quantiles. Either "normal" or "t" distributed.}

\item{ylabel}{The label for the x axis.}

\item{xlabel}{The label for the y axis.}
}
\value{
\item{data}{A dataset containing the correlational values aswell as their respective p-values.}
\item{plot}{A ggplot object of the plot.}
}
\description{
Produces a QQ plot of your data with either normal or t-distributed quantiles.
}
\examples{
data(mtcars)
qqplot(mtcars$mpg)
}
